requirements.txt
streamlit
openai
pyairtable
python-telegram-bot==20.*
pandas
